Resolved code conflicts from develop branch
