--- Cartridge_Cache/src/App.tsx
@@ -1,12 +1,0 @@
-import './App.css'
-import { Page } from './page/page'
-
-function App() {
-    return (
-        <>
-            <Page />
-        </>
-    )
-}
-
-export default App
