--- Cartridge_Cache/src/index.css
@@ -1,1 +1,0 @@
-@import "tailwindcss";
