--- Cartridge_Cache/src/page/page.tsx
@@ -1,9 +1,0 @@
-import UserProfileSection from "../components/common/user-profile/user-profile";
-
-export function Page() {
-    return(
-        <>
-            <UserProfileSection />
-        </>
-    )
-}
